rgw: s3 object lock avoids overflow in retention date
Fixes: https://tracker.ceph.com/issues/56993

(cherry picked from commit dd87a0ff845549ce94b740881c614c77c1857865)

+++ src/rgw/rgw_object_lock.cc
@@ -56,7 +56,7 @@ ceph::real_time RGWObjectLock::get_lock_until_date(const ceph::real_time& mtime)
   if (!rule_exist) {
     return ceph::real_time();
   }
-  int days = get_days();
+  int64_t days = get_days();
   if (days <= 0) {
     days = get_years()*365;
   }
